@@ -1,14 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Hosting.WindowsServices;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Hosting;
 using XExten.XCore;
 
 namespace Mily.OtherApi
@@ -17,25 +14,32 @@ namespace Mily.OtherApi
     {
         public static void Main(string[] args)
         {
-            if (!args.Contains("--Service"))
-                CreateHostBuilder(args).UseStartup<Startup>().Build().Run();
+            if (!args.Contains("--Service:"))
+                CreateHostBuilder(args).Build().Run();
             else
             {
-                //open cmd with administrator and input SC CREATE [SERVERNAME] BINPATH="[PROGRAM_URL]\*.EXE" --Service
-                args.ToEach<string>(item =>
+                args.ToEach<String>(item =>
                 {
-                    if (item.Trim().IsContainsIn("--Service:"))
+                    //open cmd with administrator and input SC CREATE [SERVERNAME] BINPATH="[PROGRAM_URL]\*.EXE" --Service:1024
+                    if (item.IsContainsIn("--Service:"))
                     {
-                        int Port = Convert.ToInt32(item.Split(":")[1].Trim());
+                        int Port = Convert.ToInt32(item.Split(":")[1]);
                         Startup.Port = Port + 1;
-                        CreateHostBuilder(args).UseKestrel()
-                        .UseContentRoot(Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName))//加载当前目录
-                        .UseUrls($"http://0.0.0.0:{Port}")//监听的IP和端口
-                        .UseStartup<Startup>().Build().RunAsService();
+                        CreateWebHostBuilder(args, Port);
                     }
                 });
             }
         }
-        public static IWebHostBuilder CreateHostBuilder(string[] args) => WebHost.CreateDefaultBuilder(args);
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+            .ConfigureWebHostDefaults(webBuilder =>
+            {
+                webBuilder.UseStartup<Startup>();
+            });
+        public static void CreateWebHostBuilder(string[] args, int Port) =>
+            WebHost.CreateDefaultBuilder(args).UseKestrel()
+                .UseContentRoot(Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName))//加载当前目录
+                .UseUrls($"http://0.0.0.0:{Port}")//监听的IP和端口
+                .UseStartup<Startup>().Build().RunAsService();
     }
 }
